+++ src/components/view_in_browser.rs
@@ -3,14 +3,15 @@ use serde::Serialize;
 use crate::components::{Component, ComponentId};
 
 #[derive(Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
 pub struct ViewInBrowser {
 	id: ComponentId,
-	text: String,
+	text: Option<String>,
 }
 
 impl ViewInBrowser {
 	pub fn new(text: Option<String>) -> Self {
-		Self { id: ComponentId::ViewInBrowser, text: text.unwrap_or_default() }
+		Self { id: ComponentId::ViewInBrowser, text }
 	}
 }